+++ okitweb/static/okit/model/js/artefacts/compartment.js
@@ -16,11 +16,11 @@ class Compartment extends OkitArtifact {
         // Configure default values
         this.compartment_id = null;
         this.name = this.generateDefaultName(okitjson.compartments.length + 1);
+        this.display_name = this.name;
+        this.description = this.name;
         // Update with any passed data
         this.merge(data);
         this.convert();
-        this.display_name = this.name;
-        this.description = this.name;
     }
 
     /*
@@ -36,7 +36,7 @@ class Compartment extends OkitArtifact {
     ** Clone Functionality
      */
     clone() {
-        return new Compartment(this, this.getOkitJson());
+        return new Compartment(JSON.clone(this), this.getOkitJson());
     }