+++ okitweb/static/okit/model/js/artefacts/security_list.js
@@ -29,7 +29,7 @@ class SecurityList extends OkitArtifact {
     ** Clone Functionality
      */
     clone() {
-        return new SecurityList(this, this.getOkitJson());
+        return new SecurityList(JSON.clone(this), this.getOkitJson());
     }
 
 
@@ -90,7 +90,10 @@ class SecurityList extends OkitArtifact {
                         "max": 22,
                         "min": 22
                     },
-                    "source_port_range": null
+                    "source_port_range": {
+                        "max": null,
+                        "min": null
+                    }
                 },
                 "udp_options": null,
                 "description": ""