+++ okitweb/static/okit/view/designer/js/artefacts/compartment.js
@@ -10,12 +10,19 @@ console.info('Loaded Designer Compartment View Javascript');
 class CompartmentView extends OkitContainerDesignerArtefactView {
     constructor(artefact=null, json_view) {
         super(artefact, json_view);
+        this.export = false;
     }
 
     get parent_id() {
         return (this.artefact.compartment_id !== null && this.artefact.compartment_id !== this.artefact.id) ? this.artefact.compartment_id : 'canvas';
     }
     get parent() {return this.getJsonView().getCompartment(this.parent_id);}
+    get children() {return [...this.json_view.getCompartments(), ...this.json_view.getVirtualCloudNetworks(),
+        ...this.json_view.getBlockStorageVolumes(), ...this.json_view.getDynamicRoutingGateways(),
+        ...this.json_view.getAutonomousDatabases(), ...this.json_view.getCustomerPremiseEquipments(),
+        ...this.json_view.getObjectStorageBuckets(), ...this.json_view.getFastConnects(),
+        ...this.json_view.getIPSecConnections(), ...this.json_view.getRemotePeeringConnections(),
+        ...this.json_view.getInstances()].filter(child => child.parent_id === this.artefact.id);}
     get minimum_dimensions() {
         if (this.isTopLevel()) {
             return {width: $(`#${this.json_view.parent_id}`).width(), height: $(`#${this.json_view.parent_id}`).height()};
@@ -29,14 +36,22 @@ class CompartmentView extends OkitContainerDesignerArtefactView {
      */
 
     isTopLevel() {
-        return this.parent ? false : true;
+        return this.parent || this.export ? false : true;
     }
 
     /*
     ** Clone Functionality
      */
     clone() {
-        return new CompartmentView(this.artefact, this.getJsonView());
+        const clone = super.clone();
+        this.cloneChildren(clone);
+        return clone;
+    }
+
+    cloneChildren(clone) {
+        for (let child of this.children) {
+            child.clone().compartment_id = clone.id;
+        }
     }
 
     /*