+++ visualiser/common/okitCommon.py
@@ -159,8 +159,28 @@ def writeAnsibleFile(ansible_file, contents):
 def writePythonFile(python_file, contents):
     logger.info('Writing Python File: {0:s}'.format(python_file))
     with closing(open(python_file, 'w')) as f:
+        for resource in contents:
+            f.write('{0:s}\n'.format(resource))
+    return
+
+
+def writeMarkdownFile(md_file, contents):
+    logger.info('Writing Markdown File: {0:s}'.format(md_file))
+    with closing(open(md_file, 'w')) as f:
+        for resource in contents:
+            f.write('{0!s:s}\n'.format(resource))
+    return
+
+
+def writeFile(filename, contents):
+    logger.info('Writing File: {0:s}'.format(filename))
+    dirname = os.path.dirname(filename)
+    if not os.path.exists(dirname):
+        os.makedirs(dirname)
+    with closing(open(filename, 'w')) as f:
         f.write('{0:s}\n'.format(contents))
     return
+
 
 def standardiseIds(json_data={}, from_char='.', to_char='-'):
     return json_data